Separate AWT and Swing apps
